Made it possible to draw ROI
This commits makes it possible to draw a region of interest in the image
both while it is paused and while it is playing. A rectangle is drawn
whenever the left mouse button is clicked inside the image and is
continuously drawn when the mouse is moved around. When the left mouse
button is released a rectangle is placed in the image with the top left
at the position where it was clicked and the bottom right where it was
released.

It is also possible to remove rectangles, pressing 'b' (as in blank)
removes all drawn rectangles and pressing the scroll button removes the
latest drawn.

Change-Id: I1a850a77d501d5de6eb3f9e8b26ef5d34aaf60d4